fixed #167 dont delete bean if it is in clipboard

+++ src/main/java/de/cismet/verdis/gui/befreiungerlaubnis/BefreiungerlaubnisTable.java
@@ -215,7 +215,9 @@ public class BefreiungerlaubnisTable extends AbstractCidsBeanTable implements Ci
     public void removeBean(final CidsBean cidsBean) {
         if (cidsBean != null) {
             try {
-                cidsBean.delete();
+                if (!Main.getInstance().getCurrentClipboard().getClipboardBeans().contains(cidsBean)) {
+                    cidsBean.delete();
+                }
             } catch (final Exception ex) {
                 LOG.error("error while removing befer", ex);
             }
